Extract advisor brief client protocols
