+++ src/app/services/advisory_client_protocols.py
@@ -628,65 +628,3 @@ class ProposalClient(Protocol):
         version_no: int,
         correlation_id: str,
     ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisorBriefAiClient(Protocol):
-    async def execute_workflow_pack(
-        self,
-        *,
-        pack_id: str,
-        version: str,
-        environment: str,
-        caller_identity_class: str,
-        workflow_surface: str | None,
-        task_request: dict[str, Any],
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_observability_runtime_status(
-        self,
-        *,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_workflow_pack_run_consumer_view(
-        self,
-        *,
-        run_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def get_workflow_pack_run_operator_profile(
-        self,
-        *,
-        run_id: str,
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def list_workflow_pack_task_flows(
-        self,
-        *,
-        correlation_id: str,
-        workflow_pack_id: str | None = None,
-        caller: str | None = None,
-        workflow_surface: str | None = None,
-        limit: int = 25,
-    ) -> tuple[int, dict[str, Any]]: ...
-
-    async def apply_workflow_pack_run_review_action(
-        self,
-        *,
-        run_id: str,
-        correlation_id: str,
-        request_payload: dict[str, Any],
-    ) -> tuple[int, dict[str, Any]]: ...
-
-
-class AdvisorBriefAdviseClient(Protocol):
-    async def get_platform_capabilities(
-        self,
-        *,
-        consumer_system: str = "lotus-gateway",
-        tenant_id: str = "default",
-        correlation_id: str,
-    ) -> tuple[int, dict[str, Any]]: ...
